+++ changelog.md
@@ -1,6 +1,7 @@
 # Changelog
 
-- * 9199b10 - 2018-06-16: Release v9.3.35 
+- * 10d61b1 - 2018-06-17: Switched WebUI to use API for login credentials 
+- * c800afc - 2018-06-16: Release v9.3.35 
 - * 2e05d78 - 2018-06-16: Added function to randomize external upnp port number Added ability to manually set external upnp port number 
 - * 8ee962a - 2018-06-16: Pre-Release v9.3.35.dev2 
 - * bcdcce3 - 2018-06-16: Pre-Release v9.3.35.dev1